Refresh options headers downloading multiple files
Headers defined in options are injected into the WebClient instance used to
download files.  This is done to support downloading files where headers
are required eg. REFERRER.  The headers can be defined and injected via
$Latest.Options.Headers.

The headers are cleared when a download is performed with DownloadFile
so where multiple files are downlaoded the headers are not available for
the second download.  This impacts when a package requries  both 32-bit
and 64-bit files to be downloaded during package construction.

To address this the headers should be injected prior to each download to
ensure that they are available.